+++ frontend/Idle/src/components/common/NormalTrimBox.jsx
@@ -1,7 +1,7 @@
 import { styled } from "styled-components";
 import { ReactComponent as ArrorRight } from "../../assets/images/arrowRight.svg";
 import { useContext } from "react";
-import { carContext } from "../../../utils/context";
+import { carContext } from "../../utils/context";
 // eslint-disable-next-line react/prop-types
 function NormalTrimBox({ purchase_rate, name, desc, price, isActive = true }) {
   const { car, setCar } = useContext(carContext);
