+++ frontend/Idle/src/pages/optionPage.jsx
@@ -1,4 +1,4 @@
-import { setClickedOptionPage } from "../../utils/constants";
+import { setClickedOptionPage } from "../utils/constants";
 function OptionPage() {
   setClickedOptionPage();
   return <div>OptionPage</div>;
