+++ frontend/Idle/src/styles/GlobalStyle.jsx
