+++ frontend/Idle/src/styles/theme.jsx
