+++ frontend/Idle/src/utils/constants.jsx
