+++ frontend/Idle/src/utils/globalCar.jsx
